pyproject.toml
[tool.poetry]
name = "simpplr"
version = "0.1.0"
description = ""
authors = ["SkMishra <sanatkumarmishra02@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
pillow = "^10.3.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
